refactor: improve database structure
