+++ src/database/db.provider.ts
@@ -1,85 +1,40 @@
 import { Injectable } from '@nestjs/common';
-import { IDatabase } from './db.interface';
-import { IUser, UserPartial } from 'src/models/user/user.interface';
+import { IUserDatabase } from './interfaces/db.interface';
+import { IUser } from './interfaces/user.interface';
 import {
   CreateUserDto,
   UpdatePasswordDto,
 } from 'src/realization/user/user.dto';
-import { generateUuid } from '../common/uuid-helper';
-import { UserNotFound, WrongOldPassword } from 'src/common/exceptions';
+import { UserDatabaseComponent } from './components/user.db';
 
 @Injectable()
-export class Database implements IDatabase {
-  private users: Map<string, UserPartial>;
+export class Database implements IUserDatabase {
+  private users: IUserDatabase;
 
   constructor() {
-    this.users = new Map();
+    this.users = new UserDatabaseComponent();
   }
 
   async getUsers(): Promise<IUser[]> {
-    const res: IUser[] = [];
-    for (const [key, value] of this.users.entries()) {
-      res.push({ id: key, ...value });
-    }
-    return new Promise((resolve) => resolve(res));
+    return await this.users.getUsers();
   }
 
   async getUser(id: string): Promise<IUser | null> {
-    const user = this.users.get(id);
-    return new Promise((resolve) =>
-      resolve(user ? { id, ...this.filterPassword(user) } : null),
-    );
+    return await this.users.getUser(id);
   }
 
   async createUsers(createUserDto: CreateUserDto): Promise<IUser> {
-    const id = generateUuid(this.users);
-    const user = {
-      login: createUserDto.login,
-      password: createUserDto.password,
-      version: 1,
-      createdAt: Date.now(),
-      updatedAt: 0,
-    };
-    this.users.set(id, user);
-    return { id, ...this.filterPassword(user) };
+    return await this.users.createUsers(createUserDto);
   }
 
   async deleteUser(id: string): Promise<void> {
-    if (this.users.has(id)) {
-      await this.users.delete(id);
-    } else {
-      throw new UserNotFound(id);
-    }
+    await this.users.deleteUser(id);
   }
 
   async updatePassword(
     id: string,
     updatePasswordDto: UpdatePasswordDto,
   ): Promise<IUser> {
-    const userData = await this.getUser(id);
-    if (!userData) {
-      throw new UserNotFound(id);
-    }
-
-    if (userData.password !== updatePasswordDto.oldPassword) {
-      throw new WrongOldPassword();
-    }
-
-    const updated = {
-      ...userData,
-      ...{
-        password: updatePasswordDto.newPassword,
-        version: ++userData.version,
-        updatedAt: Date.now(),
-      },
-    };
-    this.users.set(id, updated);
-    return new Promise((resolve) =>
-      resolve({ id, ...this.filterPassword(updated) }),
-    );
-  }
-
-  private filterPassword({ login, createdAt, updatedAt, version }) {
-    return { login, createdAt, updatedAt, version };
+    return await this.users.updatePassword(id, updatePasswordDto);
   }
 }
