+++ src/database/interfaces/artist.interface.ts
@@ -1,0 +1,5 @@
+export interface Artist {
+  id: string;
+  name: string;
+  grammy: boolean;
+}
